refactor: the browser stays in src/web, and eslint holds the line
Nine modules only the page imports (dom-utils, canvas, google-drive,
rewind-thumbnail, microphone-input, audio-utils, gamepads, speech-output,
keyboard) move under src/web, localDisc moves out of fdc.js into the web
layer with its tests, and the two remaining browser touches in core code
(a userAgent check in keymap.js already guarded for node, and the XHR the
loader uses in a browser) are written for the environment they run in.
Core src/ is now linted with node and shared globals only, so the boundary
is a rule rather than a memory; src/web, src/main.js, src/app and the
vendored GL debugger keep the browser.

Part of #1072.
