+++ src/web/config.js
@@ -1,5 +1,5 @@
 import { allModels, findModel, tubeModelFor } from "../models.js";
-import { getFilterForMode } from "../canvas.js";
+import { getFilterForMode } from "./canvas.js";
 import { AudioOutputs } from "../audio-output.js";
 
 const round = (value) => Number(value.toFixed(2));
