+++ src/web/google-drive.js
@@ -1,6 +1,6 @@
-import { discFor } from "./fdc.js";
-import { uint8ArrayToString } from "./binary.js";
-import { debounce } from "./debounce.js";
+import { discFor } from "../fdc.js";
+import { uint8ArrayToString } from "../binary.js";
+import { debounce } from "../debounce.js";
 
 const MIME_TYPE = "application/vnd.jsbeeb.disc-image";
 const CLIENT_ID = "356883185894-bhim19837nroivv18p0j25gecora60r5.apps.googleusercontent.com";
